+++ src/lib.rs
@@ -282,6 +282,7 @@ pub mod setting;
 pub mod shape;
 pub mod shared;
 pub mod signal;
+pub mod slot;
 pub mod snoop;
 pub mod sound;
 pub mod svf;